# UI/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TF26 API",
    "description": "Turkish participation-banking campaigns and live pricing. This service authenticates callers and exposes what `banks/`, `index/` and `corpus/` already do; it holds no banking logic itself.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/me/profile": {
      "get": {
        "tags": [
          "profile"
        ],
        "summary": "Get Profile",
        "description": "This user's preferences. Empty defaults when they have not answered yet.",
        "operationId": "get_profile",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "profile"
        ],
        "summary": "Put Profile",
        "description": "Write the profile. A partial body updates only the fields it carries.\n\nPUT rather than PATCH is a small lie about semantics, and a deliberate one:\nthe interview arrives in pieces as the agent asks questions, and requiring\nthe client to send the whole profile every time would mean a dropped answer\nsilently erases an earlier one.",
        "operationId": "put_profile",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/me/stats": {
      "get": {
        "tags": [
          "profile"
        ],
        "summary": "Get Stats",
        "description": "This user's own usage, for the profile overview.\n\nEight counts and two timestamps, all from tables that already exist. Written\nas separate queries rather than one join with conditional aggregates: they hit\nfour unrelated tables, each count is an index-only scan, and a single clever\nstatement here would be harder to read than the page it feeds.\n\nMessage counts go through the session join rather than a `user_id` on\n`chat_messages`, because there is no such column -- a message belongs to a\nconversation, and the conversation belongs to the user. That is the right\nshape; it just means these two counts are the only ones with a subquery.",
        "operationId": "get_stats",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatsOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/me/views": {
      "get": {
        "tags": [
          "profile"
        ],
        "summary": "List Views",
        "description": "Every dashboard this user kept, newest first.",
        "operationId": "list_views",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SavedViewOut"
                  },
                  "type": "array",
                  "title": "Response List Views"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/me/views/{slug}": {
      "put": {
        "tags": [
          "profile"
        ],
        "summary": "Put View",
        "description": "Create or replace a saved dashboard.\n\nComponent types are not checked against a catalog here. The catalog lives in\nthe frontend, and duplicating it in Python would guarantee the two drift;\nan unknown type renders as a visible placeholder tile instead -- which\nmatters, because the AI Overview page writes this JSON and a model will\neventually name a component that does not exist.",
        "operationId": "put_view",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SavedViewIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedViewOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "profile"
        ],
        "summary": "Delete View",
        "operationId": "delete_view",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/voice/transcriptions": {
      "post": {
        "tags": [
          "voice"
        ],
        "summary": "Create Voice Transcription",
        "description": "Create a Turkish transcript without sending audio off the machine.",
        "operationId": "create_voice_transcription",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_voice_transcription"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoiceTranscriptionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_create_voice_transcription": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "description": "One complete browser voice recording."
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_create_voice_transcription"
      },
...
      "Component": {
        "properties": {
          "type": {
            "type": "string",
            "maxLength": 64,
            "title": "Type"
          },
          "props": {
            "additionalProperties": true,
            "type": "object",
            "title": "Props"
          }
        },
        "type": "object",
        "required": [
          "type"
        ],
        "title": "Component",
        "description": "One tile on a dashboard.\n\n`type` names a component in the frontend catalog; `props` is whatever that\ncomponent takes. Unvalidated on purpose -- the catalog lives in TypeScript,\nand duplicating each component's prop schema in Python would guarantee the\ntwo drift. The frontend renders an unknown type as a visible placeholder."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProfileIn": {
        "properties": {
          "persona": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(customer|analyst)$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Persona"
          },
          "banks": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Banks"
          },
          "families": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Families"
          },
          "typical_amount": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Typical Amount"
          },
          "typical_term_months": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 360.0,
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Typical Term Months"
          },
          "answers": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Answers"
          }
        },
        "type": "object",
        "title": "ProfileIn",
        "description": "A profile write. Every field optional -- a partial answer is a valid answer.\n\nBank and family keys are validated against `banks.list_banks()` and\n`banks.families` in the router, not here: pydantic would have to import the\nbank layer to know them, and the error is more useful when it can name what\nis available."
      },
      "ProfileOut": {
        "properties": {
          "persona": {
            "type": "string",
            "title": "Persona"
          },
          "banks": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Banks"
          },
          "families": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Families"
          },
          "typical_amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Typical Amount"
          },
          "typical_term_months": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Typical Term Months"
          },
          "answers": {
            "additionalProperties": true,
            "type": "object",
            "title": "Answers"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          }
        },
        "type": "object",
        "required": [
          "persona",
          "banks",
          "families",
          "typical_amount",
          "typical_term_months",
          "answers",
          "completed_at"
        ],
        "title": "ProfileOut"
      },
...
      "SavedViewIn": {
        "properties": {
          "slug": {
            "type": "string",
            "pattern": "^[a-z0-9-]{1,80}$",
            "title": "Slug"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "components": {
            "items": {
              "$ref": "#/components/schemas/Component"
            },
            "type": "array",
            "title": "Components"
          },
          "generated": {
            "type": "boolean",
            "title": "Generated",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "slug",
          "title"
        ],
        "title": "SavedViewIn"
      },
      "SavedViewOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "components": {
            "items": {
              "$ref": "#/components/schemas/Component"
            },
            "type": "array",
            "title": "Components"
          },
          "generated": {
            "type": "boolean",
            "title": "Generated"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "slug",
          "title",
          "components",
          "generated",
          "updated_at"
        ],
        "title": "SavedViewOut"
      },
...
      "StatsOut": {
        "properties": {
          "chat_sessions": {
            "type": "integer",
            "title": "Chat Sessions"
          },
          "messages_sent": {
            "type": "integer",
            "title": "Messages Sent",
            "description": "Turns the user wrote."
          },
          "messages_received": {
            "type": "integer",
            "title": "Messages Received",
            "description": "Answers the assistant wrote."
          },
          "saved_tables": {
            "type": "integer",
            "title": "Saved Tables"
          },
          "automations": {
            "type": "integer",
            "title": "Automations"
          },
          "reports": {
            "type": "integer",
            "title": "Reports"
          },
          "unread_reports": {
            "type": "integer",
            "title": "Unread Reports"
          },
          "first_activity": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Activity"
          },
          "last_activity": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Activity"
          }
        },
        "type": "object",
        "required": [
          "chat_sessions",
          "messages_sent",
          "messages_received",
          "saved_tables",
          "automations",
          "reports",
          "unread_reports",
          "first_activity",
          "last_activity"
        ],
        "title": "StatsOut",
        "description": "What this user has actually done with the product.\n\nCounted at request time from rows that already exist. Nothing here is\nmaintained as a running total: a counter column would be a second source of\ntruth for a number whose first source is one indexed `COUNT`, and it would be\nwrong the first time a delete missed it.\n\n**There is deliberately no token count.** Nothing in this application records\nmodel usage -- the supervisor asks for `stream_usage` but no handler collects\nit, and the ten bank specialists' spend is unobserved entirely. A \"tokens\"\nfield would therefore have to read zero for every conversation ever held,\nwhich says something false rather than nothing."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VoiceTranscriptionOut": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          },
          "language": {
            "type": "string",
            "maxLength": 5,
            "minLength": 2,
            "title": "Language"
          },
          "processing_ms": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Processing Ms"
          }
        },
        "type": "object",
        "required": [
          "text",
          "language",
          "processing_ms"
        ],
        "title": "VoiceTranscriptionOut",
        "description": "A completed Turkish transcription."
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}